Split up contentItems/actionTypes into separate files to make it easier to find a specific action / see an overview of all actions of a specific type.
